modified:   nanos-lite/src/fs.c 	modified:   nanos-lite/src/irq.c 	modified:   nanos-lite/src/loader.c 	modified:   nanos-lite/src/syscall.c 	modified:   navy-apps/libs/libos/src/syscall.c 	modified:   navy-apps/tests/file-test/main.c

+++ navy-apps/libs/libos/src/syscall.c
@@ -61,7 +61,8 @@ int _yield() {
 }
 
 int _open(const char* path, int flags, mode_t mode) {
-  return _syscall_(SYS_open, path, flags, mode);
+  //printf(" _open in\n");
+  return _syscall_(SYS_open, (intptr_t)path, flags, mode);
 }
 
 int _write(int fd, void* buf, size_t count) {
